Proposal of multiple data fetcher in map tool
